Purpose Specification List
This is a suggested initial list of purposes and data categories. The idea is that if the consent receipt contains a set of Category-Purpose pairings drawn from predefined lists - it helps to fulfill the need for consent to specific uses.  Using the list wouldn't preclude a controller either adding their own items, or perhaps adding some added text (especially in purpose) to create greater context and specificity.